fix: tests & its workspace flow
