+++ tests/missing-account-reload/bad-one/Cargo.toml
@@ -1,15 +1,15 @@
-[package]
-name = "bad_one"
-version = "0.1.0"
-description = "Bad example - missing reload after CPI"
-edition = "2021"
-
 [workspace]
+members = [
+    "programs/*"
+]
+resolver = "2"
 
-[lib]
-crate-type = ["cdylib", "lib"]
+[profile.release]
+overflow-checks = true
+lto = "fat"
+codegen-units = 1
 
-[dependencies]
-anchor-lang = { path = "../../../lang", features = ["init-if-needed"] }
-anchor-spl = { path = "../../../spl" }
-
+[profile.release.build-override]
+opt-level = 3
+incremental = false
+codegen-units = 1
