+++ tests/missing-account-reload/bad-one/programs/bad-one/src/context/mod.rs
@@ -1,0 +1,5 @@
+pub mod initialize;
+pub mod transfer;
+
+pub use initialize::*;
+pub use transfer::*;
