+++ tests/missing-account-reload/bad-one/programs/bad-one/src/state/mod.rs
@@ -1,0 +1,4 @@
+pub mod user_account;
+
+pub use user_account::*;
+
