+++ tests/missing-account-reload/bad-one/programs/bad-one/src/state/user_account.rs
@@ -1,0 +1,11 @@
+use anchor_lang::prelude::*;
+
+#[account]
+pub struct UserAccount {
+    pub balance: u64,
+    pub authority: Pubkey,
+}
+
+impl Space for UserAccount {
+    const INIT_SPACE: usize = 8 + 8 + 32;
+}
