--- tests/missing-account-reload/bad-one/src/lib.rs
@@ -1,123 +1,0 @@
-use anchor_lang::prelude::*;
-use anchor_lang::solana_program::system_instruction;
-
-declare_id!("Fg6PaFpoGXkYsidMpWTK6W2BeZ7FEfcYkg476zPFsLnS");
-
-#[program]
-pub mod bad_one {
-    use super::*;
-
-    pub fn initialize(ctx: Context<Initialize>) -> Result<()> {
-        let user_account = &mut ctx.accounts.user_account;
-        user_account.balance = 1000;
-        user_account.authority = ctx.accounts.authority.key();
-        Ok(())
-    }
-
-    pub fn simple_transfer(ctx: Context<Transfer>, amount: u64) -> Result<()> {
-        // Direct CPI without reload - should trigger warning
-        anchor_lang::solana_program::program::invoke(
-            &system_instruction::transfer(
-                &ctx.accounts.user_account.key(),
-                &ctx.accounts.authority.key(),
-                amount,
-            ),
-            &[
-                ctx.accounts.user_account.to_account_info(),
-                ctx.accounts.authority.to_account_info(),
-            ],
-        )?;
-
-         // BAD: Access account data without reload
-        let balance = ctx.accounts.user_account.balance;
-        msg!("Balance after transfer: {}", balance);
-
-        Ok(())
-    }
-
-    pub fn transfer_with_cpi_bad(ctx: Context<Transfer>, amount: u64) -> Result<()> {
-        // Direct CPI call - should trigger warning
-        anchor_lang::solana_program::program::invoke(
-            &system_instruction::transfer(
-                &ctx.accounts.user_account.key(),
-                &ctx.accounts.authority.key(),
-                amount,
-            ),
-            &[
-                ctx.accounts.user_account.to_account_info(),
-                ctx.accounts.authority.to_account_info(),
-            ],
-        )?;
-
-         // BAD: Access account without reload after CPI
-        let balance = ctx.accounts.user_account.balance;
-        msg!("Balance: {}", balance);
-
-        Ok(())
-    }
-
-    pub fn transfer_with_multiple_cpi_bad(ctx: Context<Transfer>, amount: u64) -> Result<()> {
-        // Multiple CPI calls
-        anchor_lang::solana_program::program::invoke(
-            &system_instruction::transfer(
-                &ctx.accounts.user_account.key(),
-                &ctx.accounts.authority.key(),
-                amount,
-            ),
-            &[
-                ctx.accounts.user_account.to_account_info(),
-                ctx.accounts.authority.to_account_info(),
-            ],
-        )?;
-
-        anchor_lang::solana_program::program::invoke(
-            &system_instruction::transfer(
-                &ctx.accounts.user_account.key(),
-                &ctx.accounts.authority.key(),
-                amount,
-            ),
-            &[
-                ctx.accounts.user_account.to_account_info(),
-                ctx.accounts.authority.to_account_info(),
-            ],
-        )?;
-
-         // BAD: Access account without reload after multiple CPI
-        let balance = ctx.accounts.user_account.balance;
-        msg!("Final balance: {}", balance);
-
-        Ok(())
-    }
-}
-
-#[derive(Accounts)]
-pub struct Initialize<'info> {
-    #[account(
-        init,
-        payer = authority,
-        space = 8 + UserAccount::INIT_SPACE
-    )]
-    pub user_account: Account<'info, UserAccount>,
-    #[account(mut)]
-    pub authority: Signer<'info>,
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts)]
-pub struct Transfer<'info> {
-    #[account(mut)]
-    pub user_account: Account<'info, UserAccount>,
-    #[account(mut)]
-    pub authority: Signer<'info>,
-    pub system_program: Program<'info, System>,
-}
-
-#[account]
-pub struct UserAccount {
-    pub balance: u64,
-    pub authority: Pubkey,
-}
-
-impl Space for UserAccount {
-    const INIT_SPACE: usize = 8 + 8 + 32;
-}
