+++ tests/missing-account-reload/bad-two/programs/bad-two/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "bad-two"
+version = "0.1.0"
+description = "Bad two program"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib", "lib"]
+name = "bad_two"
+
+[features]
+default = []
+cpi = ["no-entrypoint"]
+no-entrypoint = []
+no-idl = []
+no-log-ix-name = []
+idl-build = ["anchor-lang/idl-build"]
+anchor-debug = ["anchor-lang/anchor-debug"]
+
+[dependencies]
+anchor-lang = { path = "../../../../../lang" }
+
