+++ tests/missing-account-reload/double-cpi-test/programs/double-cpi-test/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "double-cpi-test"
+version = "0.1.0"
+description = "Double CPI test program"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib", "lib"]
+name = "double_cpi_test"
+
+[features]
+default = []
+cpi = ["no-entrypoint"]
+no-entrypoint = []
+no-idl = []
+no-log-ix-name = []
+idl-build = ["anchor-lang/idl-build"]
+anchor-debug = ["anchor-lang/anchor-debug"]
+
+[dependencies]
+anchor-lang = { path = "../../../../../lang" }
+
