+++ tests/missing-account-reload/good-one/programs/good-one/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "good-one"
+version = "0.1.0"
+description = "Good one program"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib", "lib"]
+name = "good_one"
+
+[features]
+default = []
+cpi = ["no-entrypoint"]
+no-entrypoint = []
+no-idl = []
+no-log-ix-name = []
+idl-build = ["anchor-lang/idl-build"]
+anchor-debug = ["anchor-lang/anchor-debug"]
+
+[dependencies]
+anchor-lang = { path = "../../../../../lang" }
+
