+++ tests/missing-account-reload/good-two/programs/good-two/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "good-two"
+version = "0.1.0"
+description = "Good two program"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib", "lib"]
+name = "good_two"
+
+[features]
+default = []
+cpi = ["no-entrypoint"]
+no-entrypoint = []
+no-idl = []
+no-log-ix-name = []
+idl-build = ["anchor-lang/idl-build"]
+anchor-debug = ["anchor-lang/anchor-debug"]
+
+[dependencies]
+anchor-lang = { path = "../../../../../lang" }
+
