+++ tests/missing-account-reload/multi-cpi-test/programs/multi-cpi-test/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "multi-cpi-test"
+version = "0.1.0"
+description = "Multi CPI test program"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib", "lib"]
+name = "multi_cpi_test"
+
+[features]
+default = []
+cpi = ["no-entrypoint"]
+no-entrypoint = []
+no-idl = []
+no-log-ix-name = []
+idl-build = ["anchor-lang/idl-build"]
+anchor-debug = ["anchor-lang/anchor-debug"]
+
+[dependencies]
+anchor-lang = { path = "../../../../../lang" }
+
